[BE] feat : entity 필드 타입 수정

See merge request s08-webmobile1-sub2/S08P12E105!10

+++ BE/src/main/java/com/tonnybunny/domain/bunny/entity/BunnyNotiEntity.java
@@ -28,7 +28,7 @@ public class BunnyNotiEntity extends CommonEntity {
 	private LocalDate startDate;
 	private LocalDate endDate;
 
-	private Long estimatePrice;
+	private int estimatePrice;
 
 	private String startLangCode;
 	private String endLangCode;
